+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Fixed
 
+- changelog has unspecified version
+
 ### Updated
 
 ## [0.2.0-alpha.4] - 2022-07-04